403 new route path start marker (#435)
* refactored markerLayer into routePathLinkLayer

* rendering start markers of newRoutePath

* fixed mobX warning

* removed extra spaces
