+++ src/components/map/Map.tsx
@@ -12,7 +12,6 @@ import CoordinateControl from './mapControls/CoordinateControl';
 import FullscreenControl from './mapControls/FullscreenControl';
 import RouteLayer from './layers/RouteLayer';
 import NewRoutePathLayer from './layers/new/NewRoutePathLayer';
-import MarkerLayer from './layers/MarkerLayer';
 import MapLayersControl from './mapControls/MapLayersControl';
 import Toolbar from './toolbar/Toolbar';
 import EventLog from './EventLog';
@@ -160,9 +159,6 @@ class LeafletMap extends React.Component<IMapProps> {
                         fitBounds={this.fitBounds}
                     />
                     <NewRoutePathLayer />
-                    <MarkerLayer
-                        routes={routes}
-                    />
                     <PopupLayer />
                     <Control position='topleft'>
                         <Toolbar />
