--- src/components/map/layers/markerLayer.scss
@@ -1,4 +1,0 @@
-.nodeMarker {
-  display: block;
-  opacity: 0.8;
-}
