+++ src/components/map/layers/new/newRoutePathLayer.scss
@@ -35,6 +35,7 @@
 
 }
 
-
-
-
+.startMarker {
+    display: block;
+    opacity: 0.8;
+}
