+++ src/components/map/layers/routePathLinkLayer.scss
@@ -1,0 +1,4 @@
+.startMarker {
+    display: block;
+    opacity: 0.8;
+}
